DRILL-2333: Use long instead of int to store writer record count.

+++ exec/java-exec/src/main/java/org/apache/drill/exec/physical/impl/WriterRecordBatch.java
@@ -47,7 +47,7 @@ public class WriterRecordBatch extends AbstractRecordBatch<Writer> {
 
   private EventBasedRecordWriter eventBasedRecordWriter;
   private RecordWriter recordWriter;
-  private int counter = 0;
+  private long counter = 0;
   private final RecordBatch incoming;
   private boolean processed = false;
   private String fragmentUniqueId;
